pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "ragprep"
version = "0.1.0"
description = "PDF-to-Markdown app scaffold (FastAPI + htmx)"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.110",
  "jinja2>=3.1",
  "pillow>=10",
  "pypdfium2>=4.30.0",
  "pymupdf>=1.25",
  "pywebview>=5.0.0",
  "python-multipart>=0.0.9",
  "uvicorn>=0.27",
]

[dependency-groups]
dev = [
  "httpx>=0.27",
  "mypy>=1.10",
  "pytest>=8",
  "ruff>=0.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["ragprep"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
exclude = "^(dist)(/|\\\\)"
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_ignores = true
